Only use global total order.
As long as a site was in conflict, we would store the operations that it
had done such that we could roll back. This became difficult to
maintain, in regards to knowing exactly how many operations we should
keep. Using the total order and filtering it based on user seems to work
better, and is also a simplification.