Change the semantics `put'
Turns out the < operation is not transitive, which causes a problem, and
allows the ordering:

       ((del(0, f) 0 2 user 1) (ins(1, b) 0 1 user 0) (ins(0, a) 0 0 user 0))

It is not certain that we can construct a transitive relation.

For now we work around this by adding the event to the last position in
the history or is placed after any other event generated by the same
user. Then we let the event skip ahead of any events that happened
before (i.e. using the < relation).