Loosen requirements of user intent
Two simplifications are made, perhaps at cost of semantic correctness.
We don't require the "correct" character to be deleted. If we try to add
something outside the buffer bounds, add it to the end of the buffer.

This is done in order to get eventual consistency, which is a higher
ranked goal.