Remove operations that /were/ equal
We do not allow that two identical concurrent are executed (i.e. only
perform that operation once). A users local history can affect an
operation, so that they do not appear equal. We use exclusion
transformation based on a single users history to get the operation (as
if that operation was the only thing that happened at this time), and
remove duplicates based on these.