Modify put'
Still, putting an event into the total order can lead to a sequence of
operations that can't be applied on the initial state.

Using (f g h) as initial state, we can get the total order:

      (del(1, h) 1 2 user 2) (del(0, f) 0 0 user 0) (del(2,h) 0 1 user 1)

where del(1, h) and del(2,h) are semantically the same operation, but
our model does not recognize this.