Change comparison of events in the total order
We consider some events to be the same event if they are concurrent and
contain the same operation. We need a semantic notion of what the same
operation means. When comparing two events we need to reset them to an
agreed upon state, which is the lowest state token of the two events.
Both events must exclude local events, and the event with the higher
state token will need to exclude the events that happened after the
agreed upon state.

We consider to operations to be the same if they are equal after being
reset to an agreed upon state.