Send appropriate responses to incoming messages
At 54ecd7c we started saving outgoing messages at the server, but
removed responses, seeing that the scheme did not account for more than
two users.

Remember that when we receive a message (including an operation), we can
detect how many messages has been rejected at the client.

The response composed of three operations:
    - the operation sent to every other client
    - a composition of rejected operations
    - the inverse of the received operation

When we send a response, we can delete the saved outgoing messages, but
saving the response.

Here we have shown eventual consistency for 3 clients given 4 actions.