Change how events are merged
Events that contain the same operation should be merged. A counter
example was produced after 5 operations (and two clients), because we
tried to delete a character beyond the bounds of the buffer. We therefor
choose to always keep the event that contains the operation with the
smallest position. The rationale is that it which one we keep is
arbitrary, so we choose the one that won't overstep the buffer bounds.