Shrink the state space
We enumerate our users, such that they all have a unique label. The
labels in themselves are not interesting, so if we have two states that
would be equal if we changed the labels, then we should only check on of
these states.

To enforce this we restrict what users may perform an action (now any
user can perform an action if the number of remaining actions is
positive). We simply enforce that user 0 always performs the first
action, then user 0 or user 1 performs the next, then user 0, user 1 or
user 2, and so on.