Detected and fixed error in model
After disallowing insertions to surpass the buffer length (a bug in the
model), model checking was redone and found a counter example. The
counterexample needed 2 clients, 3 operations and a buffer of initial
size 3 to occur. The fix is model checked for the same initial buffer
and 3 operations and works, but the solution should be refined.